+++ src/objc/class.rs
@@ -119,7 +119,7 @@ impl Class {
     #[inline]
     #[allow(unused)] // Used by `foundation`
     pub(crate) fn alloc(&self) -> NSObject {
-        unsafe { _msg_send![self, alloc] }
+        unsafe { _msg_send_cached![self, alloc] }
     }
 
     /// Returns a reference to this class as an Objective-C object.
@@ -134,7 +134,7 @@ impl Class {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418583-respondstoselector).
     #[inline]
     pub fn responds_to_selector(&self, selector: SEL) -> bool {
-        unsafe { _msg_send![self, respondsToSelector: selector => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, respondsToSelector: selector => BOOL] != 0 }
     }
 
     /// Returns `true` if instances of this class implement or inherit a method
@@ -143,7 +143,7 @@ impl Class {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/nsobject/1418555-instancesrespondtoselector).
     #[inline]
     pub fn instances_respond_to_selector(&self, selector: SEL) -> bool {
-        unsafe { _msg_send![self, instancesRespondToSelector: selector => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, instancesRespondToSelector: selector => BOOL] != 0 }
     }
 
     /// Returns the name of this class.