+++ src/objc/ns_object.rs
@@ -29,7 +29,7 @@ impl AsRef<NSObject> for NSObject {
 impl<T: ObjectType> PartialEq<T> for NSObject {
     #[inline]
     fn eq(&self, other: &T) -> bool {
-        unsafe { _msg_send![self, isEqual: other.as_object() => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, isEqual: other.as_object() => BOOL] != 0 }
     }
 }
 
@@ -96,7 +96,7 @@ impl NSObject {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1571952-retaincount).
     #[inline]
     pub fn retain_count(&self) -> usize {
-        unsafe { _msg_send![self, retainCount] }
+        unsafe { _msg_send_cached![self, retainCount] }
     }
 
     /// Returns `true` if this object implements or inherits a method that can
@@ -105,7 +105,7 @@ impl NSObject {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418583-respondstoselector).
     #[inline]
     pub fn responds_to_selector(&self, selector: SEL) -> bool {
-        unsafe { _msg_send![self, respondsToSelector: selector => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, respondsToSelector: selector => BOOL] != 0 }
     }
 
     /// Returns `true` if this object is an instance or subclass of `class`.
@@ -113,7 +113,7 @@ impl NSObject {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418511-iskindofclass)
     #[inline]
     pub fn is_kind_of_class(&self, class: &Class) -> bool {
-        unsafe { _msg_send![self, isKindOfClass: class => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, isKindOfClass: class => BOOL] != 0 }
     }
 
     /// Returns `true` if this object is an instance of `class`.
@@ -121,7 +121,7 @@ impl NSObject {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418766-ismemberofclass)
     #[inline]
     pub fn is_member_of_class(&self, class: &Class) -> bool {
-        unsafe { _msg_send![self, isMemberOfClass: class => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, isMemberOfClass: class => BOOL] != 0 }
     }
 
     /// Returns an integer that can be used as a table address in a hash table
@@ -130,7 +130,7 @@ impl NSObject {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418859-hash).
     #[inline]
     pub fn hash(&self) -> NSUInteger {
-        unsafe { _msg_send![self, hash] }
+        unsafe { _msg_send_cached![self, hash] }
     }
 
     /// Returns a copy of this object using
@@ -139,7 +139,7 @@ impl NSObject {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/nsobject/1418807-copy).
     #[inline]
     pub fn copy(&self) -> NSObject {
-        unsafe { _msg_send![self, copy] }
+        unsafe { _msg_send_cached![self, copy] }
     }
 
     /// Returns a copy of this object using
@@ -148,6 +148,6 @@ impl NSObject {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/nsobject/1418978-mutablecopy).
     #[inline]
     pub fn mutable_copy(&self) -> NSObject {
-        unsafe { _msg_send![self, mutableCopy] }
+        unsafe { _msg_send_cached![self, mutableCopy] }
     }
 }